feat: update transaction service
